fix(cloister): isolate feedback start client
Refs: PAN-3392
